+++ .gitignore
@@ -1,4 +1,3 @@
 projet-bioinfo/
 ALL_GNPS_cleaned.mgf
-cluster_molecules4/
-smiles_invalides.txt
+Cluster.mgf
