+++ resultats_clusters/out.energy_102040_precursor_M-e.mtx.I20
@@ -1,0 +1,13 @@
+1	8	12
+%%MatrixMarket	matrix
+5	6
+9	10
+16
+2
+3
+4
+7
+11
+13
+14
+15
