+++ resultats_clusters/out.energy_13.0_precursor_M+H.mtx.I20
@@ -1,0 +1,6 @@
+3	7	9	12
+1	4
+2	8
+5	6
+10
+11
