+++ resultats_clusters/out.energy_23.0_precursor_2M-2H+Na.mtx.I20
@@ -1,0 +1,21 @@
+2	23	22	26
+33	24	27
+12	18	29
+20	25	30
+%%MatrixMarket	matrix
+4	13
+6	17
+9	10
+15	32
+1
+3
+5
+7
+8
+11
+14
+16
+19
+21
+28
+31
