+++ resultats_clusters/out.energy_27.0_precursor_M+Na.mtx.I20
@@ -1,0 +1,6 @@
+13	3	7	9	12
+1	4	6
+%%MatrixMarket	matrix
+5	8
+10	11
+2
