+++ resultats_clusters/out.energy_36.0_precursor_M-H.mtx.I20
@@ -1,0 +1,6 @@
+1	2	3	4	5	6
+%%MatrixMarket	matrix
+13	11
+7	8
+9	10
+12
