+++ resultats_clusters/out.energy_40.0_precursor_2M+Na.mtx.I20
@@ -1,0 +1,7 @@
+1	2	6	7
+14	9	11
+13	10	12
+%%MatrixMarket	matrix
+3	4
+5
+8
