+++ resultats_clusters/out.energy_43.0_precursor_2M-2H+Na.mtx.I20
@@ -1,0 +1,15 @@
+3	7	16	25	20	29
+33	1	4	13
+2	23	22	8
+24	26	27	32
+12	18	28
+%%MatrixMarket	matrix
+6	17
+9	10
+15	31
+5
+11
+14
+19
+21
+30
