+++ resultats_clusters/out.energy_45.0_precursor_M-H2O+H.mtx.I20
@@ -1,0 +1,5 @@
+3	4	5
+%%MatrixMarket	matrix
+8	1
+2	6
+7
