+++ resultats_spectres/energy_102040_precursor_M-e.mtx
@@ -1,0 +1,23 @@
+%%MatrixMarket matrix coordinate real general
+%
+16 16 20
+1 1 1
+8 1 3.975542095787911E-2
+2 2 9.999999999999999E-1
+3 3 1
+4 4 1
+5 5 1
+6 5 1
+6 6 1
+7 7 1
+8 8 1
+12 8 1.0653611370718503E-2
+9 9 1
+10 9 3.3786739812078437E-1
+10 10 1
+11 11 1
+12 12 1
+13 13 1
+14 14 1
+15 15 9.999999999999998E-1
+16 16 1
