+++ resultats_spectres/energy_13.0_precursor_M+H.mtx
@@ -1,0 +1,20 @@
+1 1 1
+4 1 9.985731285485583E-1
+2 2 1
+8 2 1
+3 3 1
+7 3 3.790783387374241E-1
+9 3 9.497587610821836E-1
+12 3 5.942127631504024E-1
+4 4 1
+5 5 1
+6 5 3.047443614657472E-1
+6 6 1
+7 7 1
+9 7 2.736386843547339E-1
+8 8 1
+9 9 1
+12 9 5.492820576806235E-1
+10 10 1
+11 11 1
+12 12 1
