+++ resultats_spectres/energy_23.0_precursor_2M-2H+Na.mtx
@@ -1,0 +1,52 @@
+%%MatrixMarket matrix coordinate real general
+%
+33 33 49
+1 1 1
+2 2 1
+23 2 1.0302746720378231E-1
+3 3 1
+4 4 1
+13 4 7.57703773904209E-1
+5 5 1
+6 6 1
+17 6 7.06135139679704E-1
+7 7 1
+8 8 1
+9 9 1
+10 9 9.934805318398455E-1
+10 10 1
+11 11 1
+12 12 1
+18 12 8.574266311572777E-2
+13 13 1
+14 14 1
+15 15 1
+32 15 7.114793311953251E-1
+16 16 1
+17 17 1
+18 18 1
+29 18 4.4417821293627824E-1
+19 19 1
+20 20 1
+23 20 2.0612223539130458E-2
+25 20 3.423393286672156E-2
+21 21 1
+22 22 1
+23 22 1.056977395784091E-1
+26 22 9.93064692574317E-1
+23 23 1
+25 23 3.966653045737608E-2
+26 23 1.0643590530281503E-1
+24 24 1
+27 24 8.793679333614539E-1
+25 25 1
+30 25 8.976015523031387E-2
+26 26 1
+27 27 9.999999999999999E-1
+33 27 7.583412970675965E-2
+28 28 1
+29 29 1
+30 30 1
+31 31 1
+32 32 1
+33 33 1
