+++ resultats_spectres/energy_26.0_precursor_M+Na.mtx
@@ -1,0 +1,12 @@
+%%MatrixMarket matrix coordinate real general
+%
+7 7 9
+1 1 1
+2 1 1.1596204496373504E-1
+2 2 1
+3 3 1
+4 4 1
+5 4 8.422315446294255E-1
+5 5 1
+6 6 1
+7 7 1
