+++ resultats_spectres/energy_27.0_precursor_M+Na.mtx
@@ -1,0 +1,31 @@
+%%MatrixMarket matrix coordinate real general
+%
+13 13 28
+1 1 1
+4 1 1.2966308861478804E-1
+6 1 1.1612604447610478E-1
+2 2 1
+3 3 1
+7 3 3.0503336866635953E-2
+9 3 5.317717700751595E-2
+4 4 1
+6 4 1.2663490927291204E-1
+5 5 9.999999999999998E-1
+7 5 1.0228333786942578E-2
+8 5 9.636271767337051E-1
+12 5 3.945579553016866E-2
+6 6 1
+7 7 1
+9 7 9.61018693723612E-1
+12 7 7.699965566809848E-2
+13 7 1.5736838223386375E-2
+8 8 1
+9 9 1
+12 9 5.947589466631317E-2
+13 9 1.4290793813605193E-2
+10 10 1
+11 10 9.981116800382395E-1
+11 11 1
+12 12 1
+13 12 4.443709720401539E-2
+13 13 9.999999999999999E-1
