+++ resultats_spectres/energy_36.0_precursor_M-H.mtx
@@ -1,0 +1,38 @@
+%%MatrixMarket matrix coordinate real general
+%
+13 13 35
+1 1 1
+2 1 8.84279523930979E-1
+3 1 8.460808421568561E-1
+4 1 9.055458663221423E-1
+5 1 8.65503550240932E-1
+6 1 9.440055954854933E-1
+2 2 1
+3 2 9.853360151524029E-1
+4 2 9.913235026627196E-1
+5 2 9.842117642633793E-1
+6 2 9.778214814425965E-1
+3 3 9.999999999999999E-1
+4 3 9.706052360583681E-1
+5 3 9.802904614164818E-1
+6 3 9.519930379059235E-1
+4 4 1
+5 4 9.589134029960371E-1
+6 4 9.902886407426865E-1
+5 5 1
+6 5 9.507715633125362E-1
+6 6 1
+7 7 1
+8 7 9.699826458638858E-1
+9 7 6.229247367053337E-2
+10 7 9.500085592817747E-2
+8 8 1
+9 8 5.881813729632271E-2
+10 8 1.255197703447087E-1
+9 9 1
+10 9 9.581858531793751E-1
+10 10 9.999999999999998E-1
+11 11 1
+13 11 9.997984546919743E-1
+12 12 1
+13 13 1
