+++ resultats_spectres/energy_37.0_precursor_M.mtx
@@ -1,0 +1,4 @@
+%%MatrixMarket matrix coordinate integer symmetric
+%
+1 1 1
+1 1 1
