+++ resultats_spectres/energy_40.0_precursor_2M+Na.mtx
@@ -1,0 +1,28 @@
+%%MatrixMarket matrix coordinate real general
+%
+14 14 25
+1 1 1
+2 1 8.835805289958371E-1
+6 1 7.934492173644171E-2
+2 2 1
+6 2 9.622480911152774E-2
+3 3 1
+4 3 9.999995492453582E-1
+4 4 1
+5 5 1
+6 6 1
+7 6 1.8160630312587592E-1
+13 6 1.2019047484026366E-2
+7 7 1
+8 8 1
+9 9 1
+11 9 1.0430236838478356E-1
+10 10 1
+12 10 6.390743581127541E-1
+13 10 4.6855493674031656E-1
+11 11 1
+14 11 7.858228004215578E-1
+12 12 1
+13 12 2.2749115894550012E-1
+13 13 1
+14 14 1
