+++ resultats_spectres/energy_43.0_precursor_2M-2H+Na.mtx
@@ -1,0 +1,64 @@
+%%MatrixMarket matrix coordinate real general
+%
+33 33 61
+1 1 1
+4 1 9.464313839045291E-2
+2 2 1
+23 2 1.6476569022765156E-1
+3 3 1
+7 3 4.4313657980024485E-2
+4 4 1
+13 4 6.329033990691362E-1
+23 4 8.485294250041786E-2
+33 4 2.3011559734405197E-2
+5 5 1
+6 6 1
+17 6 5.218286342273101E-1
+22 6 8.994110803388831E-2
+7 7 1
+16 7 1.061529385684286E-2
+25 7 2.0131170636441015E-1
+8 8 1
+23 8 6.1586825503772015E-2
+9 9 1
+10 9 9.996066541314786E-1
+10 10 9.999999999999999E-1
+11 11 1
+12 12 1
+16 12 1.0846788818134839E-2
+18 12 4.511969470022736E-1
+13 13 1
+23 13 1.2818767546133716E-1
+14 14 1
+15 15 1
+31 15 5.221262389281878E-1
+16 16 9.999999999999997E-1
+25 16 7.5077997744582734E-3
+17 17 1
+22 17 1.5032889003184731E-2
+18 18 1
+23 18 2.8006158084999486E-2
+28 18 2.67198271197087E-1
+19 19 1
+20 20 1
+23 20 7.628615119375544E-2
+25 20 2.598060199097272E-1
+21 21 1
+22 22 1
+23 22 2.894207556236169E-1
+23 23 1
+25 23 1.8165568716009442E-1
+24 24 1
+26 24 4.8067802220805905E-1
+25 25 1
+29 25 1.577054265355768E-1
+26 26 1
+27 26 7.641996338450865E-2
+27 27 1
+32 27 7.764942175002378E-2
+28 28 1
+29 29 1
+30 30 9.999999999999998E-1
+31 31 1
+32 32 1
+33 33 1
