+++ resultats_spectres/energy_44.0_precursor_M+H.mtx
@@ -1,0 +1,10 @@
+%%MatrixMarket matrix coordinate real general
+%
+4 4 7
+1 1 1
+3 1 3.82169721380333E-2
+4 1 3.695538527827143E-2
+2 2 9.999999999999998E-1
+3 3 1
+4 3 9.982157501232725E-1
+4 4 1
