+++ resultats_spectres/energy_45.0_precursor_M-H2O+H.mtx
@@ -1,0 +1,19 @@
+%%MatrixMarket matrix coordinate real general
+%
+8 8 16
+1 1 1
+8 1 1.3090857651067654E-2
+2 2 1
+3 2 9.648262079106319E-2
+4 2 1.0855104584488556E-1
+5 2 1.391123070854798E-1
+6 2 6.198985298697313E-1
+3 3 9.999999999999999E-1
+4 3 3.840817909089083E-1
+5 3 4.5154809096191184E-1
+4 4 1
+5 4 9.880550775550517E-1
+5 5 9.999999999999998E-1
+6 6 1
+7 7 1
+8 8 1
