+++ resultats_spectres/energy_53.0_precursor_M+H.mtx
@@ -1,0 +1,5 @@
+%%MatrixMarket matrix coordinate integer symmetric
+%
+2 2 2
+1 1 1
+2 2 1
